gh-279: transforms to heracles (#284)
