+++ heracles/__init__.py
@@ -69,6 +69,9 @@ __all__ = [
     "angular_power_spectra",
     "debias_cls",
     "mixing_matrices",
+    # transforms
+    "cl2corr",
+    "corr2cl",
 ]
 
 try:
@@ -138,4 +141,9 @@ from .twopoint import (
     angular_power_spectra,
     debias_cls,
     mixing_matrices,
+)
+
+from .transforms import (
+    cl2corr,
+    corr2cl,
 )
